23.06.05
(게시글 조회, 수정, 삭제 기능, 게시글 수정,삭제 시 검증 및 처리)

+++ src/main/java/com/example/blog/controller/PostController.java
@@ -7,6 +7,7 @@ import com.example.blog.domain.PostForm;
 import com.example.blog.service.CommentService;
 import com.example.blog.service.PostService;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
@@ -22,10 +24,11 @@ import java.util.List;
 
 @Controller
 @AllArgsConstructor
+@Slf4j
 public class PostController {
 
     private PostService postService;
-    private CommentService commentService;
+
     @GetMapping("/post")
     public String postForm(Model model) {
         model.addAttribute("postForm", new PostForm());
@@ -48,40 +51,65 @@ public class PostController {
     }
 
     @GetMapping("/postList")
-    public String postList(Model model) {
+    public String postList(@ModelAttribute("error") String error, Model model) {
+        if (!error.isBlank()) {
+            model.addAttribute("error", error);
+        }
         List<Post> postList = postService.getPostList();
         model.addAttribute("postList", postList);
         return "postList";
     }
 
-    @GetMapping("/post/{postSeq}")
+    @GetMapping("/postList/{postSeq}")
     public String getPost(@PathVariable Long postSeq, Model model) {
         Post findPost = postService.getPost(postSeq);
-        List<Comment> commentList = commentService.getCommentList(postSeq);
-
+        log.info("post={}", findPost);
         model.addAttribute("post", findPost);
-        model.addAttribute("comment", new Comment());
-        model.addAttribute("comments", commentList);
-        return "/post";
+        return "postOne";
     }
 
-    @GetMapping("/post/update/{postSeq}")
-    public String postUpdateForm(@PathVariable Long postSeq, HttpServletRequest request, Model model) {
+    @GetMapping("/update/{postSeq}")
+    public String postUpdateForm(@PathVariable Long postSeq, Model model, HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        // 로그인 사용자가 수정 버튼이 아닌 url로 접근했을 때, 게시 글 작성자와 일치하는 지 확인
         Post findPost = postService.getPost(postSeq);
-        model.addAttribute("post", findPost);
-        return "postList";
 
+        HttpSession session = request.getSession(false);
+        Member loginMember = (Member) session.getAttribute("loginMember");
+        if (loginMember.getMemberSeq() == findPost.getMemberSeq()) {
+            model.addAttribute("postForm", findPost);
+            return "updatePost";
+        }
+        else {
+            redirectAttributes.addFlashAttribute("error", "잘못된 접근");
+            return "redirect:/postList";
+        }
     }
 
-    @PostMapping("/post/update/{postSeq}")
-    public String postUpdate(@PathVariable Long postSeq, @ModelAttribute Post post, HttpServletRequest request, Model model) {
-        postService.updatePost(post, postSeq);
-        return "redirect:/post/{postSeq}";
+    @PostMapping("/update/{postSeq}")
+    public String postUpdate(@Validated @ModelAttribute PostForm postForm, BindingResult bindingResult, @PathVariable Long postSeq,  Model model) {
+        log.info("postForm={}", postForm.toString());
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("postForm", postForm);
+            return "updatePost";
+        }
+        postService.updatePost(postForm, postSeq);
+        return "redirect:/postList/{postSeq}";
     }
 
-    @GetMapping("/post/delete/{postSeq}")
-    public String postDelete(@PathVariable Long postSeq, HttpServletRequest request) {
-        postService.deletePost(postSeq);
-        return "redirect:/";
+    @GetMapping("/delete/{postSeq}")
+    public String postDelete(@PathVariable Long postSeq, HttpServletRequest request, RedirectAttributes redirectAttributes) {
+        Post post = postService.getPost(postSeq);
+
+        HttpSession session = request.getSession(false);
+        Member loginMember = (Member) session.getAttribute("loginMember");
+        if (loginMember.getMemberSeq() == post.getMemberSeq()) {
+            postService.deletePost(postSeq);
+            return "redirect:/";
+        } else {
+            redirectAttributes.addFlashAttribute("error", "잘못된 접근");
+            return "redirect:/postList";
+        }
+
+
     }
 }
